backend/blog/requirements.txt
asgiref==3.5.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-taggit==4.0.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
packaging==21.3
Pillow==9.2.0
pip-review==1.2.0
pipreqs==0.4.13
psycopg2-binary==2.9.4
pyparsing==3.0.9
pytz==2022.5
requests==2.31.0
sqlparse==0.4.3
urllib3==2.0.3
whitenoise==6.2.0
yarg==0.1.9
